Refactored and added new svg code
