+++ exmol/__init__.py
@@ -2,3 +2,4 @@ from .version import __version__
 from . import stoned
 from .exmol import *
 from .stoned import sanitize_smiles
+from .plot_utils import rewrite_svg, mpl2svg, moldiff
